+++ src/shape/index.ts
@@ -1,3 +1,4 @@
 export * from './asyncShape.js'
 export * from './different.js'
+export * from './identical.js'
 export * from './shape.js'